pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
